requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
anthropic==0.34.2
pdfplumber==0.11.4
python-docx==1.1.2
sqlalchemy==2.0.35
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
alembic==1.13.3
pytest==8.3.3
httpx==0.27.2
